Finished implementation of core-LoG method. Also implemented helpful visualization function 'paintBlobs'.

+++ src/GaussianBlob.h
@@ -17,7 +17,7 @@ class GaussianBlob
      */
 public:
     // GaussianBlob has no default constructor.
-    GaussianBlob(int x1, int x2, double sigma, uchar log_value){
+    GaussianBlob(int x1, int x2, double sigma, float log_value){
         /**
         * Constructor.
         */
@@ -44,7 +44,7 @@ public:
         return sqrt(2) * _sigma;
     }
 
-    [[nodiscard]] uchar get_log_value() const{
+    [[nodiscard]] float get_log_value() const{
         return _log_value;
     }
 
@@ -56,7 +56,7 @@ private:
     int _x1;         // Vertical location.
     int _x2;         // Horizontal location.
     double _sigma;     // Standard deviation.
-    uchar _log_value;  // The value of the scale-normalized Laplacian at that blob.
+    float _log_value;  // The value of the scale-normalized Laplacian at that blob.
                         // This can be seen as measure for the strength of the blob in the image.
 };
